WIP: Refactor of CPUCompute and stand alone cpu-render-test
