+++ tools/cpu-render-test/cpu-render-test-tool.vcxproj
@@ -19,11 +19,10 @@
     </ProjectConfiguration>
   </ItemGroup>
   <PropertyGroup Label="Globals">
-    <ProjectGuid>{61F7EB00-7281-4BF3-9470-7C2EA92620C3}</ProjectGuid>
+    <ProjectGuid>{5701695E-7324-4B4D-977A-8D56C2A041B1}</ProjectGuid>
     <IgnoreWarnCompileDuplicatedFilename>true</IgnoreWarnCompileDuplicatedFilename>
     <Keyword>Win32Proj</Keyword>
-    <RootNamespace>render-test-tool</RootNamespace>
-    <WindowsTargetPlatformVersion>10.0.14393.0</WindowsTargetPlatformVersion>
+    <RootNamespace>cpu-render-test-tool</RootNamespace>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'" Label="Configuration">
@@ -69,29 +68,29 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
     <LinkIncremental>true</LinkIncremental>
     <OutDir>..\..\bin\windows-x86\debug\</OutDir>
-    <IntDir>..\..\intermediate\windows-x86\debug\render-test-tool\</IntDir>
-    <TargetName>render-test-tool</TargetName>
+    <IntDir>..\..\intermediate\windows-x86\debug\cpu-render-test-tool\</IntDir>
+    <TargetName>cpu-render-test-tool</TargetName>
     <TargetExt>.dll</TargetExt>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|x64'">
     <LinkIncremental>true</LinkIncremental>
     <OutDir>..\..\bin\windows-x64\debug\</OutDir>
-    <IntDir>..\..\intermediate\windows-x64\debug\render-test-tool\</IntDir>
-    <TargetName>render-test-tool</TargetName>
+    <IntDir>..\..\intermediate\windows-x64\debug\cpu-render-test-tool\</IntDir>
+    <TargetName>cpu-render-test-tool</TargetName>
     <TargetExt>.dll</TargetExt>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <LinkIncremental>false</LinkIncremental>
     <OutDir>..\..\bin\windows-x86\release\</OutDir>
-    <IntDir>..\..\intermediate\windows-x86\release\render-test-tool\</IntDir>
-    <TargetName>render-test-tool</TargetName>
+    <IntDir>..\..\intermediate\windows-x86\release\cpu-render-test-tool\</IntDir>
+    <TargetName>cpu-render-test-tool</TargetName>
     <TargetExt>.dll</TargetExt>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <LinkIncremental>false</LinkIncremental>
     <OutDir>..\..\bin\windows-x64\release\</OutDir>
-    <IntDir>..\..\intermediate\windows-x64\release\render-test-tool\</IntDir>
-    <TargetName>render-test-tool</TargetName>
+    <IntDir>..\..\intermediate\windows-x64\release\cpu-render-test-tool\</IntDir>
+    <TargetName>cpu-render-test-tool</TargetName>
     <TargetExt>.dll</TargetExt>
   </PropertyGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
@@ -107,7 +106,7 @@
     <Link>
       <SubSystem>Windows</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <ImportLibrary>..\..\bin\windows-x86\debug\render-test-tool.lib</ImportLibrary>
+      <ImportLibrary>..\..\bin\windows-x86\debug\cpu-render-test-tool.lib</ImportLibrary>
     </Link>
     <PostBuildEvent>
       <Command>"$(SolutionDir)tools\copy-hlsl-libs.bat" "$(WindowsSdkDir)Redist/D3D/x86/" "../../bin/windows-x86/debug/"</Command>
@@ -126,7 +125,7 @@
     <Link>
       <SubSystem>Windows</SubSystem>
       <GenerateDebugInformation>true</GenerateDebugInformation>
-      <ImportLibrary>..\..\bin\windows-x64\debug\render-test-tool.lib</ImportLibrary>
+      <ImportLibrary>..\..\bin\windows-x64\debug\cpu-render-test-tool.lib</ImportLibrary>
     </Link>
     <PostBuildEvent>
       <Command>"$(SolutionDir)tools\copy-hlsl-libs.bat" "$(WindowsSdkDir)Redist/D3D/x64/" "../../bin/windows-x64/debug/"</Command>
@@ -149,7 +148,7 @@
       <SubSystem>Windows</SubSystem>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
-      <ImportLibrary>..\..\bin\windows-x86\release\render-test-tool.lib</ImportLibrary>
+      <ImportLibrary>..\..\bin\windows-x86\release\cpu-render-test-tool.lib</ImportLibrary>
     </Link>
     <PostBuildEvent>
       <Command>"$(SolutionDir)tools\copy-hlsl-libs.bat" "$(WindowsSdkDir)Redist/D3D/x86/" "../../bin/windows-x86/release/"</Command>
@@ -172,30 +171,30 @@
       <SubSystem>Windows</SubSystem>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
-      <ImportLibrary>..\..\bin\windows-x64\release\render-test-tool.lib</ImportLibrary>
+      <ImportLibrary>..\..\bin\windows-x64\release\cpu-render-test-tool.lib</ImportLibrary>
     </Link>
     <PostBuildEvent>
       <Command>"$(SolutionDir)tools\copy-hlsl-libs.bat" "$(WindowsSdkDir)Redist/D3D/x64/" "../../bin/windows-x64/release/"</Command>
     </PostBuildEvent>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClInclude Include="cpu-compute-util.h" />
-    <ClInclude Include="cpu-memory-binding.h" />
-    <ClInclude Include="options.h" />
-    <ClInclude Include="png-serialize-util.h" />
-    <ClInclude Include="shader-input-layout.h" />
-    <ClInclude Include="shader-renderer-util.h" />
-    <ClInclude Include="slang-support.h" />
+    <ClInclude Include="..\render-test\cpu-compute-util.h" />
+    <ClInclude Include="..\render-test\cpu-memory-binding.h" />
+    <ClInclude Include="..\render-test\options.h" />
+    <ClInclude Include="..\render-test\png-serialize-util.h" />
+    <ClInclude Include="..\render-test\shader-input-layout.h" />
+    <ClInclude Include="..\render-test\shader-renderer-util.h" />
+    <ClInclude Include="..\render-test\slang-support.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="cpu-compute-util.cpp" />
-    <ClCompile Include="cpu-memory-binding.cpp" />
-    <ClCompile Include="options.cpp" />
-    <ClCompile Include="png-serialize-util.cpp" />
-    <ClCompile Include="render-test-main.cpp" />
-    <ClCompile Include="shader-input-layout.cpp" />
-    <ClCompile Include="shader-renderer-util.cpp" />
-    <ClCompile Include="slang-support.cpp" />
+    <ClCompile Include="..\render-test\cpu-compute-util.cpp" />
+    <ClCompile Include="..\render-test\cpu-memory-binding.cpp" />
+    <ClCompile Include="..\render-test\cpu-render-test-main.cpp" />
+    <ClCompile Include="..\render-test\options.cpp" />
+    <ClCompile Include="..\render-test\png-serialize-util.cpp" />
+    <ClCompile Include="..\render-test\shader-input-layout.cpp" />
+    <ClCompile Include="..\render-test\shader-renderer-util.cpp" />
+    <ClCompile Include="..\render-test\slang-support.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\source\core\core.vcxproj">
