+++ tools/render-test/slang-support.h
@@ -6,6 +6,7 @@
 #include <slang.h>
 
 #include "shader-input-layout.h"
+#include "options.h"
 
 namespace renderer_test {
 
@@ -52,7 +53,19 @@ struct ShaderCompilerUtil
         ShaderProgram::Desc desc;
         SlangCompileRequest* request = nullptr;
         SlangSession* session = nullptr;
+
     };
+
+    struct OutputAndLayout
+    {
+        Output output;
+        ShaderInputLayout layout;
+        Slang::String sourcePath;
+    };
+
+    static SlangResult compileWithLayout(SlangSession* session, const Slang::String& sourcePath, Options::ShaderProgramType shaderType, const ShaderCompilerUtil::Input& input, OutputAndLayout& output);
+
+    static SlangResult readSource(const Slang::String& inSourcePath, List<char>& outSourceText);
 
     static SlangResult compileProgram(SlangSession* session, const Input& input, const ShaderCompileRequest& request, Output& out);
 };
